requirements.txt
selenium==4.17.2
pytest==8.0.0
webdriver-manager==4.0.2
pytest-html==4.1.1
allure-pytest==2.13.2
python-dotenv==1.0.0
flake8
